Added the snapshot functionnality
Users can now click on the upper right button, select the resolution of
their choice (defaults to the actual resolution of the world) and take a
snapshot. A preview of the shot is shown in the menu box and the user
can then download the snapshot in full resolution, or click on the
preview to delete the image and try again.