Finalized info bubbles when clicking on organs
Info bubbles pop up when clicking on an organ. For the moment, only
the organ number is displayed, but this can be changed easily when there
will be more info on each organ. The bubbles might need to follow the 3D
model later.

A future PR will add the feature to grab bubbles to move them and keep
them saved somewhere on the viewport.